% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppk_models.R
\name{error_model_comb2}
\alias{error_model_comb2}
\title{Residuel error model combined 2}
\usage{
error_model_comb2(f, xi)
}
\arguments{
\item{f}{Numeric vector, output of a pharmacokinetic model}

\item{xi}{Numeric vector of the coefficients for the
residual error model}
}
\value{
Numeric vector, residual error
}
\description{
Residual error model combined 2.
}
\details{
Implements the following function:
\code{g <- sqrt(xi[1]^2 + xi[2]^2*f^2)}
}
