% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_amikacin_2cpt_Burdet2015}
\alias{mod_amikacin_2cpt_Burdet2015}
\title{Two-compartment model of IV amikacin (Burdet 2015)}
\format{
A  list of six objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{theta}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{sigma}{The estimates of the parameters of the residual error model}
}
}
\source{
\doi{10.1007/s00228-014-1766-y}
}
\usage{
mod_amikacin_2cpt_Burdet2015
}
\description{
Population pharmacokinetics model from a study of single-dose amikacin
in critically ill patients with suspected ventilator-associated pneumonia
}
\details{
Covariates included in the model:
\describe{
\item{CLCREAT4H}{4-h creatinine clearance in ml/min}
\item{TBW}{Total body weight in kg}
\item{PoverF}{PaO2/FIO2 ratio in mmHg}
}
}
\keyword{datasets}
