% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_amox_oral_1cpt_fictional}
\alias{mod_amox_oral_1cpt_fictional}
\title{Fictionnal oral amoxicillin model for test purposes}
\format{
A list of five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\source{
In-house model created specifically for posologyr
}
\usage{
mod_amox_oral_1cpt_fictional
}
\description{
A fictional one-compartment population pharmacokinetic
model of oral amoxicillin, with non-saturable oral
absorption, and non-zero covariance between the
inter-individual variability of V and Cl.
}
\keyword{datasets}
