% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_amoxicillin_oral_1cpt_fictional}
\alias{mod_amoxicillin_oral_1cpt_fictional}
\title{FICTIONAL oral amoxicillin model for test purpose}
\format{
A  list of six objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{theta}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{sigma}{The estimates of the parameters of the residual error model}
}
}
\source{
In-house model created specifically for posologyr
}
\usage{
mod_amoxicillin_oral_1cpt_fictional
}
\description{
A FICTIONAL one-compartment population pharmacokinetic
model of oral amoxicillin, with non-saturable oral
absorption, and non-zero covariance between the
inter-individual variability of V and Cl.
}
\keyword{datasets}
