% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_piperacillin_2cpt_Roberts2010}
\alias{mod_piperacillin_2cpt_Roberts2010}
\title{Two-compartment model of IV piperacillin (Roberts 2010)}
\format{
A  list of seven objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{theta}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters of inter-individual variability (IIV)}
\item{pi_matrix}{A named square variance-covariance matrix of the
population parameters of intra-individual variability (aka
inter-occasion variability, IOV)}
\item{covariates}{A character vector of the covariates of
the model}
\item{sigma}{The estimates of the parameters of the residual error model}
}
}
\source{
\doi{10.1016/j.ijantimicag.2009.10.008}
}
\usage{
mod_piperacillin_2cpt_Roberts2010
}
\description{
Population pharmacokinetics model from a study of piperacillin
by continuous or intermittent dosing in critically ill patients
with sepsis
}
\details{
Covariate included in the model:
\describe{
\item{TBW}{Total body weight in kg}
}
}
\keyword{datasets}
