% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_tobramycin_2cpt_fictional}
\alias{mod_tobramycin_2cpt_fictional}
\title{Fictional IV tobramycin model for test purpose}
\format{
A  list of six objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{psi}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{xi}{The estimates of the parameters of the residual error model}
}
}
\source{
\url{https://www.page-meeting.org/pdf_assets/1954-2017_05_01_poster_Tobramycin.pdf}
}
\usage{
mod_tobramycin_2cpt_fictional
}
\description{
A fictional bicompartmental population pharmacokinetic
model of intravenous tobramycin.
}
\keyword{datasets}
