% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_vancomycin_2cpt_Goti2018}
\alias{mod_vancomycin_2cpt_Goti2018}
\title{Two-compartment model of IV vancomycin (Goti 2018)}
\format{
A  list of six objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{theta}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{xi}{The estimates of the parameters of the residual error model}
}
}
\source{
\url{https//doi.org/10.1097/FTD.0000000000000490}
}
\usage{
mod_vancomycin_2cpt_Goti2018
}
\description{
Population pharmacokinetics model from a study of hospitalized
patients with and without hemodialysis
}
\details{
Covariates included in the model:
\describe{
\item{CLCREAT}{Creatinine clearance in ml/min calculated by the
Cockroft-Gault formula}
\item{WT}{Body weight in kg}
\item{DIAL}{Hemodialysis status: 0 in the case of non-dialysis
subjects, 1 in the case of dialysis subjects}
}
}
\keyword{datasets}
