% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing_optim.R
\name{poso_dose_auc}
\alias{poso_dose_auc}
\title{Estimate the dose needed to reach a target area under the concentration-time
curve (AUC)}
\usage{
poso_dose_auc(
  dat = NULL,
  prior_model = NULL,
  tdm = FALSE,
  time_auc,
  time_dose = NULL,
  cmt_dose = 1,
  target_auc,
  estim_method = "map",
  nocb = FALSE,
  p = NULL,
  greater_than = TRUE,
  starting_time = 0,
  interdose_interval = NULL,
  add_dose = NULL,
  duration = 0,
  starting_dose = 100,
  indiv_param = NULL
)
}
\arguments{
\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/rxode2 event records.}

\item{prior_model}{A \code{posologyr} prior population pharmacokinetics
model, a list of six objects.}

\item{tdm}{A boolean. If \code{TRUE}: estimates the optimal dose for a selected
target auc over a selected duration following the events from \code{dat}, and
using Maximum A Posteriori estimation. Setting \code{tdm} to \code{TRUE} causes the
following to occur:
\itemize{
\item the \code{time_dose} argument is required and is used as the starting point
for the AUC calculation instead of \code{starting_time};
\item the arguments \code{estim_method}, \code{p}, \code{greater_than}, \code{interdose_interval},
\code{add_dose}, \code{indiv_param} and \code{starting_time} are ignored.
}}

\item{time_auc}{Numeric. A duration. The target AUC is computed from
\code{starting_time} to \code{starting_time} + \code{time_auc}.
When \code{tdm} is set to \code{TRUE} the target AUC is computed from \code{time_dose} to
\code{time_dose} + \code{time_auc} instead.}

\item{time_dose}{Numeric. Time when the dose is to be given. Only used and
mandatory, when \code{tdm} is set to \code{TRUE}.}

\item{cmt_dose}{Character or numeric. The compartment in which the dose is
to be administered. Must match one of the compartments in the prior model.
Defaults to 1.}

\item{target_auc}{Numeric. The target AUC.}

\item{estim_method}{A character string. An estimation method to be used for
the individual parameters. The default method "map" is the Maximum A
Posteriori estimation, the method "prior" simulates from the prior
population model, and "sir" uses the Sequential Importance Resampling
algorithm to estimate the a posteriori distribution of the individual
parameters. This argument is ignored if \code{indiv_param} is provided, or if
\code{tdm} is set to \code{TRUE}.}

\item{nocb}{A boolean. for time-varying covariates: the next observation
carried backward (nocb) interpolation style, similar to NONMEM.  If
\code{FALSE}, the last observation carried forward (locf) style will be used.
Defaults to \code{FALSE}.}

\item{p}{Numeric. The proportion of the distribution of AUC to consider for
the optimization. Mandatory for \code{estim_method=sir}. This argument is
ignored if \code{tdm} is set to \code{TRUE}.}

\item{greater_than}{A boolean. If \code{TRUE}: targets a dose leading to a
proportion \code{p} of the AUCs to be greater than \code{target_auc}. Respectively,
lower if \code{FALSE}. This argument is ignored if \code{tdm} is set to \code{TRUE}.}

\item{starting_time}{Numeric. First point in time of the AUC, for multiple
dose regimen. The default is zero. This argument is ignored if \code{tdm} is
set to \code{TRUE}, and \code{time_dose} is used as a starting point instead.}

\item{interdose_interval}{Numeric. Time for the interdose interval for
multiple dose regimen. Must be provided when add_dose is used. This
argument is ignored if \code{tdm} is set to \code{TRUE}.}

\item{add_dose}{Numeric. Additional doses administered at inter-dose interval
after the first dose. Optional. This argument is ignored if \code{tdm} is set
to \code{TRUE}.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}

\item{starting_dose}{Numeric. Starting dose for the optimization
algorithm.}

\item{indiv_param}{Optional. A set of individual parameters : THETA,
estimates of ETA, and covariates. This argument is ignored if \code{tdm} is
set to \code{TRUE}.}
}
\value{
A list containing the following components:
\describe{
\item{dose}{Numeric. An optimal dose for the selected target AUC.}
\item{type_of_estimate}{Character string. The type of estimate of the
individual parameters. Either a point estimate, or a distribution.}
\item{auc_estimate}{A vector of numeric estimates of the AUC. Either a
single value (for a point estimate of ETA), or a distribution.}
\item{indiv_param}{A \code{data.frame}. The set of individual parameters used
for the determination of the optimal dose : THETA, estimates of ETA, and
covariates}
}
}
\description{
estimates the dose needed to reach a target area under the concentration-time
curve (AUC) given a population pharmacokinetic model, a set of individual
parameters, and a target AUC.
}
\examples{
rxode2::setRxThreads(2L) # limit the number of threads

# model

}
