% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_dose_ctime}
\alias{poso_dose_ctime}
\title{Estimate the optimal dose for a selected target concentration at a
selected time point}
\usage{
poso_dose_ctime(
  object = NULL,
  param_map = NULL,
  time_c = NULL,
  starting_dose = 100,
  interdose_interval = NULL,
  time_span = NULL,
  duration = NULL,
  target_conc = NULL
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called.}

\item{time_c}{Numeric. Point in time for which the dose is to be
optimized.}

\item{starting_dose}{Numeric. Starting dose for the optimization
algorithm.}

\item{interdose_interval}{Numeric. Time for the interdose interval
for multiple dose regimen. Must be provided when time_span is used.}

\item{time_span}{Numeric. Period over which the doses are
administered for multiple dose regimen. Optional.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}

\item{target_conc}{Numeric. Target concentration.}
}
\value{
A numeric optimal dose to reach the target concentration
at the selected point in time.
}
\description{
Estimates the optimal dose for a selected target concentration at a
selected time point given a population pharmacokinetic model, a set
of individual parameters, a selected point in time, and a target
concentration.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the optimal dose to reach a concentration of 80 mg/l
# one hour after starting the 30-minutes infusion
poso_dose_ctime(patient01_tobra,time_c=1,duration=0.5,target_conc=80)

}
