% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{poso_plot_map}
\alias{poso_plot_map}
\title{Plot the individual Maximum A Posteriori prediction over individual
observations}
\usage{
poso_plot_map(
  prior_model = prior_ppk_model,
  dat = dat_posology,
  param_psi_map = NULL,
  solved_model = solved_ppk_model
)
}
\arguments{
\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of seven elements (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{param_psi_map}{A vector of individual parameters. May be omitted
if a \code{solved_model} is provided, in which case the
\code{\link{poso_estim_map}}
function will be called}

\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object, created
with the prior RxODE structural population pharmacokinetics model and
the prior typical values of the population parameters from the
\code{prior_model}, using \code{dat} as the event record. May be omitted
if a vector of individual parameters \code{param_psi_map} is provided}
}
\value{
A \code{ggplot2} plot of the individual Maximum A Posteriori
time-concentration curve, using the individual covariates, over the
individual observations
}
\description{
Plots the individual Maximum A Posteriori prediction (or Empirical
Bayes Estimate), using the individual covariates, over the individual
observations
}
\details{
The default values of the arguments \code{solved_model}, \code{prior_model} and
\code{dat} correspond to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
seven elements:
\describe{
\item{$description}{A brief description of the population
pharmacokinetics model}
\item{$reference}{A named character vector. Bibliographic reference
of the model (DOI)}
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
plot_indiv_map()

}
