% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{poso_plot_pop}
\alias{poso_plot_pop}
\title{Plot typical population model prediction over individual
observations}
\usage{
poso_plot_pop(prior_model = prior_ppk_model, dat = dat_posology, covar = NULL)
}
\arguments{
\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of seven elements (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{covar}{a named vector of the individual covariates. If omitted,
the covariates from the individual event record \code{dat} will be
used}
}
\value{
A \code{ggplot2} plot of the typical population time-concentration curve,
using the individual covariates, over the individual observations
}
\description{
Plots the typical population pharmacokinetic model, using the
the individual covariates, over the individual observations
}
\details{
The default values of the arguments \code{prior_model} and \code{dat} correspond
to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# plot the typical population model prediction over individual
# observations
poso_plot_pop()

}
