% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppk_models.R
\name{posologyr}
\alias{posologyr}
\title{Creates a posologyr list from a prior model and an individual event
record}
\usage{
posologyr(prior_model = NULL, dat = NULL)
}
\arguments{
\item{prior_model}{A \code{posologyr} prior population pharmacokinetics model, a
list of six objects.}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records.

\code{posologyr} will check the validity of the compiled RxODE
model. If \code{prior_model$ppk_model$isValid()} returns \code{FALSE},
\code{posologyr} will call \code{\link[RxODE]{RxODE}}
to recompile the model before solving it using \code{prior_model}
and \code{dat}.}
}
\value{
A list of eight objects: the posologyr prior population
pharmacokinetics model given as \code{prior_model} parameter (6 objects),
the individual event record (\code{tdm_data}) given as \code{dat} parameter,
and the solved model (\code{solved_ppk_model}).
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{theta}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{sigma}{The estimates of the parameters of the residual error model}
\item{tdm_data}{A dataframe. The individual subject dataset
given as \code{dat} parameter}
\item{solved_ppk_model}{An \code{\link[RxODE]{rxSolve}} solve object,
created with \code{prior_ppk_model} and using \code{dat} as the event record.}
}
}
\description{
Creates a list for a \code{posologyr} prior model, an individual event
record, and an \code{\link[RxODE]{rxSolve}} solve object,
created from the prior ppk model and the individual event record.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)

}
