% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxui.R
\name{posologyr_error_lines}
\alias{posologyr_error_lines}
\alias{posologyr_error_lines.norm}
\alias{posologyr_error_lines.t}
\alias{posologyr_error_lines.default}
\alias{posologyr_error_lines.rxUi}
\title{This creates the posologyr error lines from a rxui model}
\usage{
posologyr_error_lines(line)

\method{posologyr_error_lines}{norm}(line)

\method{posologyr_error_lines}{t}(line)

\method{posologyr_error_lines}{default}(line)

\method{posologyr_error_lines}{rxUi}(line)
}
\arguments{
\item{line}{line to parse}
}
\value{
error lines for posology
}
\description{
This creates the posologyr error lines from a rxui model
}
\author{
Matthew L. Fidler
}
\keyword{internal}
