% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppk_models.R
\name{error_model_comb1}
\alias{error_model_comb1}
\title{Residual error model combined 1}
\usage{
error_model_comb1(f, xi)
}
\arguments{
\item{f}{Numeric vector, output of a pharmacokinetic model}

\item{xi}{Numeric vector of the coefficients for the
residual error model}
}
\value{
Numeric vector, residual error
}
\description{
Residual error model combined 1. Constant error model
if no proportional coefficient is provided. Proportional
error model if no constant (or additive) error
coefficient is provided.
}
\details{
Implements the following function:
\code{g <- xi[1] + xi[2]*f}
}
