% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_error.R
\name{error_model_comb2}
\alias{error_model_comb2}
\title{Residual error model combined 2}
\usage{
error_model_comb2(f, sigma)
}
\arguments{
\item{f}{Numeric vector, output of a pharmacokinetic model}

\item{sigma}{Numeric vector of the coefficients for the
residual error model}
}
\value{
Numeric vector, residual error
}
\description{
Residual error model combined 2.
}
\details{
Implements the following function:
\code{g <- sqrt(sigma[1]^2 + sigma[2]^2*f^2)}
}
\keyword{internal}
