% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppk_models.R
\name{load_ppk_model}
\alias{load_ppk_model}
\title{Easy loading of model and event record for posologyr}
\usage{
load_ppk_model(prior_model = NULL, dat = NULL)
}
\arguments{
\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of six objects (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}
}
\value{
This function is invoked for its side effect, assigning
3 objects to the environment of the function call:
\describe{
\item{prior_ppk_model}{The posologyr prior population pharmacokinetics
model given as \code{prior_model} parameter}
\item{dat_posologyr}{A dataframe. The individual subject dataset
given as \code{dat} parameter}
\item{solved_ppk_model}{An \code{\link[RxODE]{rxSolve}} solve object,
created with \code{prior_ppk_model} and using \code{dat} as the event record.}
}
}
\description{
Creates, or renames, and loads in the global environment the
objects needed for posologyr functions: prior model, solved
model, and individual event record. The objects are assigned
with names corresponding to the defaults of posologyr functions.
}
\details{
The posologyr prior population pharmacokinetics model is a list of
six objects:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{$error_model}{A function of the residual error model}
\item{$psi}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{$Omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}

\code{load_ppk_model} will check the validity of the compiled RxODE
model. If \code{prior_model$ppk_model$isValid()} returns \code{FALSE},
\code{load_ppk_model} will call \code{\link[RxODE]{RxODE}}
to recompile the model before solving it using \code{prior_model}
and \code{dat}.
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)

}
