% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_tobramycin_2cpt_fictional}
\alias{mod_tobramycin_2cpt_fictional}
\title{Fictionnal IV tobramycin model for test purposes}
\format{
A list of five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\source{
\url{https://www.page-meeting.org/pdf_assets/1954-2017_05_01_poster_Tobramycin.pdf}
}
\usage{
mod_tobramycin_2cpt_fictional
}
\description{
A fictional bicompartmental population pharmacokinetic
model of intravenous tobramycin.
}
\keyword{datasets}
