% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_dose_auc}
\alias{poso_dose_auc}
\title{Estimate the optimal dose for a selected target area under the
time-concentration curve (AUC)}
\usage{
poso_dose_auc(
  object = NULL,
  param_map = NULL,
  time_auc = NULL,
  starting_dose = 100,
  duration = NULL,
  target_auc = NULL
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called}

\item{time_auc}{a numeric last point in time of the AUC for which the dose
is to be optimized. The AUC is computed from 0 to \code{time_auc}}

\item{starting_dose}{numeric starting dose for the optimization
algorithm}

\item{duration}{a numeric duration of infusion, for zero-order
administrations}

\item{target_auc}{a numeric target AUC}
}
\value{
A numeric optimal dose to reach the target AUC
}
\description{
Estimates the optimal dose for a selected target area under the
time-concentration curve (AUC) given a population pharmacokinetic
model, a set of individual parameters, and a target AUC.
}
\details{

}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the optimal dose to reach an AUC(0-12h) of 45 h.mg/l
poso_dose_auc(patient01_tobra,time_auc=12,target_auc=45)

}
