% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_dose_auc}
\alias{poso_dose_auc}
\title{Estimate the optimal dose for a selected target area under the
time-concentration curve (AUC)}
\usage{
poso_dose_auc(
  object,
  param_map = NULL,
  adapt = FALSE,
  time_auc,
  starting_time = 0,
  starting_dose = 100,
  interdose_interval = NULL,
  add_dose = NULL,
  duration = NULL,
  target_auc
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called.}

\item{adapt}{A boolean. If \code{param_map} is omitted, should the estimation
be performed with the adaptive MAP method (as opposed to the
standard MAP)? A column \code{AMS} is required in the patient record
to define the segments for the adaptive MAP approach.}

\item{time_auc}{Numeric. Last point in time of the AUC for which the dose
is to be optimized. The AUC is computed from 0 to \code{time_auc}.}

\item{starting_time}{Numeric. First point in time of the AUC, for multiple
dose regimen. The default is zero.}

\item{starting_dose}{Numeric. Starting dose for the optimization
algorithm.}

\item{interdose_interval}{Numeric. Time for the interdose interval
for multiple dose regimen. Must be provided when add_dose is used.}

\item{add_dose}{Numeric. Additional doses administered at
inter-dose interval after the first dose. Optional.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}

\item{target_auc}{Numeric. Target AUC}
}
\value{
A numeric optimal dose to reach the target AUC.
}
\description{
Estimates the optimal dose for a selected target area under the
time-concentration curve (AUC) given a population pharmacokinetic
model, a set of individual parameters, and a target AUC.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the optimal dose to reach an AUC(0-12h) of 45 h.mg/l
poso_dose_auc(patient01_tobra,time_auc=12,target_auc=45)

}
