% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslgyr_dosing_optim.R
\name{poso_dose_auc}
\alias{poso_dose_auc}
\title{Estimate the optimal dose for a selected target area under the
time-concentration curve (AUC)}
\usage{
poso_dose_auc(
  object,
  time_auc,
  target_auc,
  estim_method = "map",
  p = NULL,
  greater_than = TRUE,
  starting_time = 0,
  interdose_interval = NULL,
  add_dose = NULL,
  duration = NULL,
  starting_dose = 100,
  indiv_param = NULL
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{time_auc}{Numeric. The target AUC is computed from 0 to \code{time_auc}.}

\item{target_auc}{Numeric. The target AUC.}

\item{estim_method}{A character string. An estimation method to be used for
the individual parameters. The default method "map" is the Maximum A
Posteriori estimation, the method "prior" simulates from the prior
population model, and "sir" uses the Sequential Importance Resampling
algorithm to estimate the a posteriori distribution of the individual
parameters. This argument is ignored if \code{indiv_param} is provided.}

\item{p}{Numeric. The proportion of the distribution of AUC to consider for
the optimization. Mandatory for \code{estim_method=sir}.}

\item{greater_than}{A boolean. If \code{TRUE}: targets a dose leading to a
proportion \code{p} of the AUCs to be greater than \code{target_auc}. Respectively,
lower if \code{FALSE}.}

\item{starting_time}{Numeric. First point in time of the AUC, for multiple
dose regimen. The default is zero.}

\item{interdose_interval}{Numeric. Time for the interdose interval for
multiple dose regimen. Must be provided when add_dose is used.}

\item{add_dose}{Numeric. Additional doses administered at
inter-dose interval after the first dose. Optional.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}

\item{starting_dose}{Numeric. Starting dose for the optimization
algorithm.}

\item{indiv_param}{Optional. A set of individual parameters : THETA,
estimates of ETA, and covariates.}
}
\value{
A list containing the following components:
\describe{
\item{dose}{Numeric. An optimal dose for the selected target AUC.}
\item{type_of_estimate}{Character string. The type of estimate of the
individual parameters. Either a point estimate, or a distribution.}
\item{auc_estimate}{A vector of numeric estimates of the AUC. Either a
single value (for a point estimate of ETA), or a distribution.}
\item{indiv_param}{A \code{data.frame}. The set of individual parameters used
for the determination of the optimal dose : THETA, estimates of ETA, and
covariates}
}
}
\description{
Estimates the optimal dose for a selected target area under the
time-concentration curve (AUC) given a population pharmacokinetic
model, a set of individual parameters, and a target AUC.
}
\examples{
# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv\%*\%sigma\%*\%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the optimal dose to reach an AUC(0-12h) of 45 h.mg/l
poso_dose_auc(patient01,time_auc=12,target_auc=45)

}
