% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_dose_ctime}
\alias{poso_dose_ctime}
\title{Estimate the optimal dose for a selected target concentration at a
selected time point}
\usage{
poso_dose_ctime(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posologyr,
  param_map = NULL,
  time_c = NULL,
  starting_dose = 100,
  duration = NULL,
  target_conc = NULL
)
}
\arguments{
\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object, created
with the prior RxODE structural population pharmacokinetics model and
the prior typical values of the population parameters from the
\code{prior_model}, using \code{dat} as the event record. May be omitted
if a vector of individual parameters \code{param_map} is provided}

\item{prior_model}{A posologyr prior population pharmacokinetics model,
a list of six objects (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records. May be omitted
if a vector of individual parameters \code{param_map} is provided}

\item{param_map}{A vector of individual parameters. May be omitted
if a \code{solved_model} and an individual event record \code{dat} are
provided, in which case the \code{\link{poso_estim_map}} function will be
called}

\item{time_c}{a numeric point in time for which the dose is to be
optimized}

\item{starting_dose}{numeric starting dose for the optimization
algorithm}

\item{duration}{a numeric duration of infusion, for zero-order
administrations}

\item{target_conc}{a numeric target concentration}
}
\value{
A numeric optimal dose to reach the target concentration
at the selected point in time
}
\description{
Estimates the optimal dose for a selected target concentration at a
selected time point given a population pharmacokinetic model, a set
of individual parameters, a selected point in time, and a target
concentration.
}
\details{
The default values of the arguments \code{solved_model}, \code{prior_model} and
\code{dat} correspond to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
six objects:
\describe{
\item{ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{error_model}{A function of the residual error model}
\item{psi}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{covariates}{A character vector of the covariates of
the model}
\item{xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# estimate the optimal dose to reach a concentration of 80 mg/l
# one hour after starting the infusion
poso_dose_ctime(time_c=1,target_conc=80)

}
