% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_dose_ctime}
\alias{poso_dose_ctime}
\title{Estimate the optimal dose for a selected target concentration at a
selected time point}
\usage{
poso_dose_ctime(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posology,
  param_psi_map = NULL,
  covar = NULL,
  time_c = NULL,
  starting_dose = 100,
  duration = NULL,
  target_ctime = NULL
)
}
\arguments{
\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object, created
with the prior RxODE structural population pharmacokinetics model and
the prior typical values of the population parameters from the
\code{prior_model}, using \code{dat} as the event record. May be omitted
if a vector of individual parameters \code{param_psi_map} is provided}

\item{prior_model}{A posologyr prior population pharmacokinetics model,
a list of seven elements (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{param_psi_map}{A vector of individual parameters. May be omitted
if a \code{solved_model} and an individual event record \code{dat} are
provided, in which case the \code{\link{poso_estim_map}} function will be
called}

\item{covar}{a named vector of the individual covariates. May be
omitted if an individual event record \code{dat} is provided}

\item{time_c}{a numeric point in time for which the dose is to be
optimized.}

\item{starting_dose}{numeric starting dose for the optimization
algorithm}

\item{duration}{a numeric duration of infusion, for zero-order
administrations}

\item{target_ctime}{a numeric target concentration}
}
\value{
A numeric optimal dose to reach the target concentration
at the selected point in time
}
\description{
Estimates the optimal dose for a selected target concentration at a
selected time point given a population pharmacokinetic model, a set
of individual parameters, a selected point in time, and a target
concentration.
}
\details{
The default values of the arguments \code{solved_model}, \code{prior_model} and
\code{dat} correspond to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
seven elements:
\describe{
\item{$description}{A brief description of the population
pharmacokinetics model}
\item{$reference}{A named character vector. Bibliographic reference
of the model (DOI)}
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
poso_dose_ctime(time_c=1,target_ctime=80)

}
