% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslgyr_param_estim.R
\name{poso_estim_map}
\alias{poso_estim_map}
\title{Estimate the Maximum A Posteriori individual parameters}
\usage{
poso_estim_map(
  object,
  adapt = FALSE,
  return_model = TRUE,
  return_ofv = FALSE,
  return_AMS_models = FALSE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{adapt}{A boolean. Should the estimation be performed with the
adaptive MAP method (as opposed to the standard MAP)? A column
\code{AMS} is required in the patient record to define the segments for
the adaptive MAP approach.}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}. If \code{adapt=TRUE}, the model is solved using the
parameters estimated on the last segment.}

\item{return_ofv}{A boolean. Returns a the Objective Function Value (OFV)
if set to \code{TRUE}. Always considered \code{FALSE} if \code{adapt=TRUE}.}

\item{return_AMS_models}{A boolean. Returns a RxODE model using the estimated
ETAs for each Adaptive MAP Segment (AMS) if set to \code{TRUE}. Ignored if
\code{adapt=FALSE}.}
}
\value{
A named list consisting of one or more of the following elements
depending on the input parameters of the function: \verb{$eta} a named vector
of the MAP estimates of the individual values of ETA, \verb{$model} an RxODE
model using the estimated ETAs, \code{AMS_models} a list of RxODE models, one for
each Adaptive MAP Segment (AMS).
}
\description{
Estimates the Maximum A Posteriori (MAP) individual parameters,
also known as Empirical Bayes Estimates (EBE).
}
\examples{
# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv\%*\%sigma\%*\%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map(patient01)

}
