% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_estim_map}
\alias{poso_estim_map}
\title{Estimate the Maximum A Posteriori individual parameters}
\usage{
poso_estim_map(
  object = NULL,
  return_model = TRUE,
  return_fim = FALSE,
  return_rse = FALSE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}.}

\item{return_fim}{A boolean. Returns the Fisher Information Matrix
(FIM) if set to \code{TRUE}.}

\item{return_rse}{A boolean. Returns the relative standard errors
(RSE) of the MAP estimates of ETA if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a named vector of the MAP estimates
of the individual values of ETA.
If \code{return_model} is set to \code{TRUE}, a list of the MAP estimates of the individual
values of ETA, and a RxODE model using the estimated ETAs.
}
\description{
Estimates the Maximum A Posteriori (MAP) individual parameters,
also known as Empirical Bayes Estimates (EBE).
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map(patient01_tobra)

}
