% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_estim_map}
\alias{poso_estim_map}
\title{Estimate the Maximum A Posteriori individual parameters}
\usage{
poso_estim_map(
  object,
  adapt = FALSE,
  return_model = TRUE,
  return_fim = FALSE,
  return_rse = FALSE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{adapt}{A boolean. Should the estimation be performed with the
adaptive MAP method (as opposed to the standard MAP)? A column
\code{AMS} is required in the patient record to define the segments for
the adaptive MAP approach.}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}.}

\item{return_fim}{A boolean. Returns the Fisher Information Matrix
(FIM) if set to \code{TRUE}.}

\item{return_rse}{A boolean. Returns the relative standard errors
(RSE) of the MAP estimates of ETA if set to \code{TRUE}.}
}
\value{
A named list consisting of one or more of the following elements
depending on the input parameters of the function: \verb{$eta} a named vector
of the MAP estimates of the individual values of ETA, \verb{$model} an RxODE
model using the estimated ETAs, \verb{$fim} the Fisher information matrix,
\verb{$rse} a named vector of RSEs of the MAP estimates of ETAs.
}
\description{
Estimates the Maximum A Posteriori (MAP) individual parameters,
also known as Empirical Bayes Estimates (EBE).
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map(patient01_tobra)

}
