% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslgyr_param_estim.R
\name{poso_estim_map}
\alias{poso_estim_map}
\title{Estimate the Maximum A Posteriori individual parameters}
\usage{
poso_estim_map(
  object,
  adapt = FALSE,
  return_model = TRUE,
  return_ofv = FALSE,
  return_AMS_models = FALSE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{adapt}{A boolean. Should the estimation be performed with the
adaptive MAP method (as opposed to the standard MAP)? A column
\code{AMS} is required in the patient record to define the segments for
the adaptive MAP approach.}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}. If \code{adapt=TRUE}, the model is solved using the
parameters estimated on the last segment.}

\item{return_ofv}{A boolean. Returns a the Objective Function Value (OFV)
if set to \code{TRUE}. Always considered \code{FALSE} if \code{adapt=TRUE}.}

\item{return_AMS_models}{A boolean. Returns a RxODE model using the estimated
ETAs for each Adaptive MAP Segment (AMS) if set to \code{TRUE}. Ignored if
\code{adapt=FALSE}.}
}
\value{
A named list consisting of one or more of the following elements
depending on the input parameters of the function: \verb{$eta} a named vector
of the MAP estimates of the individual values of ETA, \verb{$model} an RxODE
model using the estimated ETAs, \code{AMS_models} a list of RxODE models, one for
each Adaptive MAP Segment (AMS).
}
\description{
Estimates the Maximum A Posteriori (MAP) individual parameters,
also known as Empirical Bayes Estimates (EBE).
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map(patient01_tobra)

}
