% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslgyr_param_estim.R
\name{poso_estim_sir}
\alias{poso_estim_sir}
\title{Estimate the posterior distribution of individual parameters by SIR}
\usage{
poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{n_sample}{Number of samples from the S-step}

\item{n_resample}{Number of samples from the R-step}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a list of one element: a
dataframe \verb{$eta} of ETAs from the posterior distribution, estimated by
Sequential Importance Resampling.
If \code{return_model} is set to \code{TRUE}, a list of the dataframe of the posterior
distribution of ETA, and a RxODE model using the estimated distributions of ETAs.
}
\description{
Estimates the posterior distribution of individual parameters by
Sequential Importance Resampling (SIR)
}
\examples{
# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv\%*\%sigma\%*\%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)

}
