% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_estim_sir}
\alias{poso_estim_sir}
\title{Estimate the posterior distribution of individual parameters by SIR}
\usage{
poso_estim_sir(
  object,
  n_sample = 1e+05,
  n_resample = 1000,
  return_model = TRUE
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{n_sample}{Number of samples from the S-step}

\item{n_resample}{Number of samples from the R-step}

\item{return_model}{A boolean. Returns a RxODE model using the estimated
ETAs if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a list of one element: a
dataframe \verb{$eta} of ETAs from the posterior distribution, estimated by
Sequential Importance Resampling.
If \code{return_model} is set to \code{TRUE}, a list of the dataframe of the posterior
distribution of ETA, and a RxODE model using the estimated distributions of ETAs.
}
\description{
Estimates the posterior distribution of individual parameters by
Sequential Importance Resampling (SIR)
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

}
