% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_inter_cmin}
\alias{poso_inter_cmin}
\title{Estimate the optimal inter-dose interval for a given dose and a
selected target trough concentration}
\usage{
poso_inter_cmin(
  object,
  param_map = NULL,
  adapt = FALSE,
  dose,
  starting_interval = 12,
  add_dose = 10,
  duration = NULL,
  target_cmin
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called.}

\item{adapt}{A boolean. If \code{param_map} is omitted, should the estimation
be performed with the adaptive MAP method (as opposed to the
standard MAP)? A column \code{AMS} is required in the patient record
to define the segments for the adaptive MAP approach.}

\item{dose}{Numeric. The dose given.}

\item{starting_interval}{Numeric. Starting inter-dose interval for
the optimization algorithm.}

\item{add_dose}{Numeric. Additional doses administered at
inter-dose interval after the first dose.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}

\item{target_cmin}{Numeric. Target trough concentration (Cmin).}
}
\value{
A numeric inter-dose interval to reach the target trough
concentration before each dosing of a multiple dose regimen.
}
\description{
Estimates the optimal inter-dose interval for a selected target
trough concentration (Cmin), given a dose, a population
pharmacokinetic model, a set of individual parameters, and a
target concentration.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the optimal interval to reach a cmin of of 2.5 mg/l
# before each administration
poso_inter_cmin(patient01_tobra,dose=1500,duration=0.5,target_cmin=2.5)

}
