% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_simu_pop}
\alias{poso_simu_pop}
\title{Estimate the prior distribution of population parameters}
\usage{
poso_simu_pop(prior_model = prior_ppk_model, n_simul = 1000)
}
\arguments{
\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of seven elements (see 'Details' for the description of the
object)}

\item{n_simul}{An integer, the number of simulations to be run}
}
\value{
A list of 2: a dataframe of the simulated population
pharmacokinetic parameters, and a vector of the prior typical
values of the population pharmacokinetic parameters
}
\description{
Estimates the prior distribution of population parameters by Monte Carlo
simulations
}
\details{
The posologyr prior population pharmacokinetics model is a list of
five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# estimate the prior distribution of population parameters
poso_simu_pop()

}
