% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_simu_pop}
\alias{poso_simu_pop}
\title{Estimate the prior distribution of population parameters}
\usage{
poso_simu_pop(prior_model = prior_ppk_model, n_simul = 1000)
}
\arguments{
\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of seven elements (see 'Details' for the description of the
object)}

\item{n_simul}{An integer, the number of simulations to be run}
}
\value{
A list of 2: a dataframe of the simulated population
pharmacokinetic parameters, and a vector of the prior typical
values of the population pharmacokinetic parameters
}
\description{
Estimates the prior distribution of population parameters by Monte Carlo
simulations
}
\details{
The posologyr prior population pharmacokinetics model is a list of
seven elements:
\describe{
\item{$description}{A brief description of the population
pharmacokinetics model}
\item{$reference}{A named character vector. Bibliographic reference
of the model (DOI)}
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
poso_simu_pop()

}
