% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_estim.R
\name{poso_simu_pop}
\alias{poso_simu_pop}
\title{Estimate the prior distribution of population parameters}
\usage{
poso_simu_pop(object, n_simul = 1000, return_model = TRUE)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{n_simul}{An integer, the number of simulations to be run. For \code{n_simul =0}, all ETAs are set to 0.}

\item{return_model}{A boolean. Returns a rxode2 model using the simulated
ETAs if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a list of one element: a
dataframe \verb{$eta} of the individual values of ETA.
If \code{return_model} is set to \code{TRUE}, a list of the dataframe of the
individual values of ETA, and a rxode2 model using the simulated ETAs.
}
\description{
Estimates the prior distribution of population parameters by Monte Carlo
simulations
}
\examples{
# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv\%*\%sigma\%*\%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)

}
