% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_simu_pop}
\alias{poso_simu_pop}
\title{Estimate the prior distribution of population parameters}
\usage{
poso_simu_pop(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posologyr,
  n_simul = 1000,
  return_model = TRUE
)
}
\arguments{
\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object, created
with the prior RxODE structural population pharmacokinetics model and the
prior estimates of the population parameters from the
\code{prior_model}, using \code{dat} as the event record}

\item{prior_model}{A posologyr prior population pharmacokinetics model, a
list of six objects (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{n_simul}{An integer, the number of simulations to be run}

\item{return_model}{A boolean. Returns a RxODE model using the simulated
ETAs if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a dataframe of the
individual values of ETA.
If \code{return_model} is set to \code{TRUE}, a list of the dataframe of the
individual values of ETA, and a RxODE model using the simulated ETAs.
}
\description{
Estimates the prior distribution of population parameters by Monte Carlo
simulations
}
\details{
The posologyr prior population pharmacokinetics model is a list of
six objects:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{$error_model}{A function of the residual error model}
\item{$psi}{A named vector of the population estimates of the
fixed effects parameters (called THETAs, following NONMEM
terminology)}
\item{$Omega}{A named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# estimate the prior distribution of population parameters
poso_simu_pop()

}
