% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{poso_simu_pop}
\alias{poso_simu_pop}
\title{Estimate the prior distribution of population parameters}
\usage{
poso_simu_pop(object = NULL, n_simul = 1000, return_model = TRUE)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{n_simul}{An integer, the number of simulations to be run.}

\item{return_model}{A boolean. Returns a RxODE model using the simulated
ETAs if set to \code{TRUE}.}
}
\value{
If \code{return_model} is set to \code{FALSE}, a dataframe of the
individual values of ETA.
If \code{return_model} is set to \code{TRUE}, a list of the dataframe of the
individual values of ETA, and a RxODE model using the simulated ETAs.
}
\description{
Estimates the prior distribution of population parameters by Monte Carlo
simulations
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

}
