% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_time_cmin}
\alias{poso_time_cmin}
\title{Predict time to a selected trough concentration}
\usage{
poso_time_cmin(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posology,
  param_psi_map = NULL,
  covar = NULL,
  from = 0.2,
  last_time = 72,
  dose = NULL,
  duration = NULL,
  target_cmin = NULL
)
}
\arguments{
\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object,
created with the prior RxODE structural population pharmacokinetics
model and the prior typical values of the population parameters
from the \code{prior_model}, using \code{dat} as the event record. May be omitted
if a vector of individual parameters \code{param_psi_map} is provided}

\item{prior_model}{A posologyr prior population pharmacokinetics model,
a list of seven elements (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{param_psi_map}{A vector of individual parameters. May be omitted
if a \code{solved_model} and an individual event record \code{dat} are
provided, in which case the \code{\link{poso_estim_map}} function
will be called}

\item{covar}{a named vector of the individual covariates. May be
omitted if an individual event record \code{dat} is provided}

\item{from}{a numeric starting time for the simulation of the
individual time-concentration profile. The default value is
0.2}

\item{last_time}{a numeric ending time for the simulation of the
individual time-concentration profile. The default value is
72}

\item{dose}{a numeric dose administered}

\item{duration}{a numeric duration of infusion, for zero-order
administrations}

\item{target_cmin}{a numeric target trough concentration}
}
\value{
A numeric time to the selected trough concentration, from the
time of administration
}
\description{
Predicts the time needed to reach a selected trough concentration
(Cmin) given a population pharmacokinetic model, a set of individual
parameters, a dose, and a target Cmin.
}
\details{
The default values of the arguments \code{solved_model}, \code{prior_model} and
\code{dat} correspond to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with no inter-individual
variability, or residual error model}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 3. \code{name}: a character vector of the names
of the population pharmacokinetc paramters, \code{reference}: a named
vector of the prior typical value of the population paramaters,
\code{Omega}: a square variance-covariance matrix of the population
parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
poso_time_cmin(dose=2500,target_cmin=2.5)

}
