% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_time_cmin}
\alias{poso_time_cmin}
\title{Predict time to a selected trough concentration}
\usage{
poso_time_cmin(
  object = NULL,
  param_map = NULL,
  from = 0.2,
  last_time = 72,
  dose = NULL,
  duration = NULL,
  target_cmin = NULL
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called.}

\item{from}{a numeric starting time for the simulation of the
individual time-concentration profile. The default value is
0.2}

\item{last_time}{a numeric ending time for the simulation of the
individual time-concentration profile. The default value is
72.}

\item{dose}{a numeric dose administered.}

\item{duration}{a numeric duration of infusion, for zero-order
administrations.}

\item{target_cmin}{a numeric target trough concentration.}
}
\value{
A numeric time to the selected trough concentration, from the
time of administration
}
\description{
Predicts the time needed to reach a selected trough concentration
(Cmin) given a population pharmacokinetic model, a set of individual
parameters, a dose, and a target Cmin.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# predict the time needed to reach a concentration of 2.5 mg/l
# after the administration of a 2500 mg dose
poso_time_cmin(patient01_tobra,dose=2500,target_cmin=2.5)

}
