% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_time_cmin}
\alias{poso_time_cmin}
\title{Predict time to a selected trough concentration}
\usage{
poso_time_cmin(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posologyr,
  param_map = NULL,
  from = 0.2,
  last_time = 72,
  dose = NULL,
  duration = NULL,
  target_cmin = NULL
)
}
\arguments{
\item{solved_model}{An \code{\link[RxODE]{rxSolve}} solve object,
created with the prior RxODE structural population pharmacokinetics
model and the prior typical values of the population parameters
from the \code{prior_model}, using \code{dat} as the event record. May be omitted
if a vector of individual parameters \code{param_map} is provided}

\item{prior_model}{A posologyr prior population pharmacokinetics model,
a list of five elements (see 'Details' for the description of the
object)}

\item{dat}{Dataframe. An individual subject dataset following the
structure of NONMEM/RxODE event records}

\item{param_map}{A vector of individual parameters. May be omitted
if a \code{solved_model} and an individual event record \code{dat} are
provided, in which case the \code{\link{poso_estim_map}} function
will be called}

\item{from}{a numeric starting time for the simulation of the
individual time-concentration profile. The default value is
0.2}

\item{last_time}{a numeric ending time for the simulation of the
individual time-concentration profile. The default value is
72}

\item{dose}{a numeric dose administered}

\item{duration}{a numeric duration of infusion, for zero-order
administrations}

\item{target_cmin}{a numeric target trough concentration}
}
\value{
A numeric time to the selected trough concentration, from the
time of administration
}
\description{
Predicts the time needed to reach a selected trough concentration
(Cmin) given a population pharmacokinetic model, a set of individual
parameters, a dose, and a target Cmin.
}
\details{
The default values of the arguments \code{solved_model}, \code{prior_model} and
\code{dat} correspond to the objects created by the convenience function
\code{\link{load_ppk_model}}

The posologyr prior population pharmacokinetics model is a list of
five elements:
\describe{
\item{$ppk_model}{A RxODE model implementing the structural
population pharmacokinetics model with the individual model
(i.e. the model of inter-individual variability) and the
covariates}
\item{$error_model}{A function of the residual error model}
\item{$pk_prior}{A list of 2. \code{psi}: a named
vector of the population estimates of the fixed effects
parameters (called THETAs, following NONMEM terminology),
\code{Omega}: a named square variance-covariance matrix of the
population parameters inter-individual variability}
\item{$covariates}{A character vector of the covariates of
the model}
\item{$xi}{The estimates of the parameters of the residual error model}
}
}
\examples{
# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# predict the time needed to reach a concentration of 2.5 mg/l
# after the administration of a 2500 mg dose
poso_time_cmin(dose=2500,target_cmin=2.5)

}
