% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{poso_time_cmin}
\alias{poso_time_cmin}
\title{Predict time to a selected trough concentration}
\usage{
poso_time_cmin(
  object,
  dose,
  target_cmin,
  param_map = NULL,
  adapt = FALSE,
  from = 0.2,
  last_time = 72,
  add_dose = NULL,
  interdose_interval = NULL,
  duration = NULL
)
}
\arguments{
\item{object}{A posologyr list, created by the \code{\link{posologyr}}
function.}

\item{dose}{Numeric. Dose administered.}

\item{target_cmin}{Numeric. Target trough concentration (Cmin).}

\item{param_map}{A vector of individual parameters. May be omitted,
in which case the \code{\link{poso_estim_map}} function
will be called.}

\item{adapt}{A boolean. If \code{param_map} is omitted, should the estimation
be performed with the adaptive MAP method (as opposed to the
standard MAP)? A column \code{AMS} is required in the patient record
to define the segments for the adaptive MAP approach.}

\item{from}{Numeric. Starting time for the simulation of the
individual time-concentration profile. The default value is
0.2}

\item{last_time}{Numeric. Ending time for the simulation of the
individual time-concentration profile. The default value is
72.}

\item{add_dose}{Numeric. Additional doses administered at
inter-dose interval after the first dose. Optional.}

\item{interdose_interval}{Numeric. Time for the inter-dose interval
for multiple dose regimen. Must be provided when add_dose is used.}

\item{duration}{Numeric. Duration of infusion, for zero-order
administrations.}
}
\value{
A numeric time to the selected trough concentration, from the
time of administration.
}
\description{
Predicts the time needed to reach a selected trough concentration
(Cmin) given a population pharmacokinetic model, a set of individual
parameters, a dose, and a target Cmin.
}
\examples{
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# predict the time needed to reach a concentration of 2.5 mg/l
# after the administration of a 2500 mg dose over a 30 minutes
# infusion
poso_time_cmin(patient01_tobra,dose=2500,duration=0.5,target_cmin=2.5)

}
